Added readme and draft code which was tested (recompiling) 10 times with no deadlock
