+++ ProjectY/projy.c
@@ -1,0 +1,219 @@
+// AUTHORS: Elena, Nicole, Tim
+// Semaphore solution to chain smoker problem
+// C program, must be compiled with gcc -pthread projy.c
+
+#include <stdio.h>
+#include <pthread.h>
+#include <semaphore.h>
+#include <unistd.h>
+#include <time.h>
+#include <stdlib.h>
+#include <unistd.h>
+
+/* Main players - Agent, Smokers with tobacco, paper, match */
+sem_t agentSem, tobaccoSem, paperSem, matchSem;
+
+/* "helpers" - required */
+int isTobacco = 0, isPaper = 0, isMatch = 0; // Supplies on the table
+sem_t tobacco, paper, match; // Pushers (move supplies from agent to usert)
+sem_t mutex; //Pushers must have exclusive access to supplies on the table
+
+
+// Tobacco and Paper agent
+void *agentOne(){
+  for (int i = 0; i < 6; i++) {
+    sleep((rand()%200)/1000); //Sleep thread for up to 200 ms to mix things up
+    sem_wait(&agentSem); //Wait until previous agent deal is finished
+    printf("      Tobacco and paper generated by agent.\n");
+    sem_post(&tobacco); //wake up a tobacco pusher
+    sem_post(&paper); //wake up a paper pusher
+  }
+}
+
+// Paper and Match agent
+void *agentTwo(){
+  for (int i = 0; i < 6; i++) {
+    sleep((rand()%200)/1000); //Sleep thread for up to 200 ms to mix things up
+    sem_wait(&agentSem); //Wait until previous agent deal is finished
+    printf("      Paper and match generated by agent.\n");
+    sem_post(&paper); //wake up a paper pusher
+    sem_post(&match); //wake up a match pusher
+  }
+}
+
+// Match and Tobacco agent
+void *agentThree(){
+  for (int i = 0; i < 6; i++) {
+    sleep((rand()%200)/1000); //Sleep thread for up to 200 ms to mix things up
+    sem_wait(&agentSem); //Wait until previous agent deal is finished
+    printf("      Match and tobacco generated by agent.\n");
+    sem_post(&match); //wake up a match pusher
+    sem_post(&tobacco); //wake up a tobacco pusher
+  }
+}
+
+// Pusher thread which reacts to tobacco
+void *tobaccoPusher(){
+  for (int i = 0; i < 12; i++) {
+    sem_wait(&tobacco);  // Wakes up when an agent provides tobacco
+    sem_wait(&mutex); // Must have exclusive access to table from other pushers
+    if (isPaper) {
+        // Paper and Tobacco are on the table - wake up matches
+        isPaper = 0;
+        printf("   Tobacco pusher sees paper, wakes match-holder\n");
+        sem_post(&matchSem);
+    }
+    else if (isMatch) {
+        // Matches and Tobacco are on the table - wake up paper
+        isMatch = 0;
+        printf("   Tobacco pusher sees match, wakes paper-holder\n");
+        sem_post(&paperSem);
+    }
+    else { //Nothing else on the table - put down tobacco
+        isTobacco = 1;
+        printf("   Tobacco pusher sees no other supplies, puts tobacco on table.\n");
+    }
+    // give back mutex
+    sem_post(&mutex);
+  }
+}
+
+// Pusher thread which reacts to matches
+void *matchPusher(){
+  for (int i = 0; i < 12; i++) {
+    sem_wait(&match);  // Wakes up when an agent provides matches
+    sem_wait(&mutex); // Must have exclusive access to table from other pushers
+    if (isPaper) {
+        // Paper and matches are on the table - wake up tobacco
+        isPaper = 0;
+        printf("   Match pusher sees paper, wakes tobacco-holder\n");
+        sem_post(&tobaccoSem);
+    }
+    else if (isTobacco) {
+        // Matches and Tobacco are on the table - wake up paper
+        isTobacco = 0;
+        printf("   Match pusher sees tobacco, wakes paper-holder\n");
+        sem_post(&paperSem);
+    }
+    else { //Nothing else on the table - put down matches
+        isMatch = 1;
+        printf("   Match pusher sees no other supplies, puts matches on table.\n");
+    }
+    // give back mutex
+    sem_post(&mutex);
+  }
+}
+
+// Pusher thread which reacts to paper
+void *paperPusher(){
+  for (int i = 0; i < 12; i++) {
+    sem_wait(&paper);  // Wakes up when an agent provides paper
+    sem_wait(&mutex); // Must have exclusive access to table from other pushers
+    if (isTobacco) {
+        // Paper and Tobacco are on the table - wake up matches
+        isTobacco = 0;
+        printf("   Paper pusher sees tobacco, wakes match-holder\n");
+        sem_post(&matchSem);
+    }
+    else if (isMatch) {
+        // Matches and paper are on the table - wake up tobacco
+        isMatch = 0;
+        printf("   Paper pusher sees match, wakes tobacco-holder\n");
+        sem_post(&tobaccoSem);
+    }
+    else { //Nothing else on the table - put down paper
+      isPaper = 1;
+      printf("   Paper pusher sees no other supplies, puts paper on table.\n");
+    }
+    // give back mutex
+    sem_post(&mutex);
+  }
+}
+
+
+void *tobaccoSmoker(void* arg){
+  int num = *((int*)arg); //unpack arg
+  for (int i = 0; i < 3; i++) {
+    sem_wait(&tobaccoSem);
+    sleep((rand()%50)/1000); //Make cigarette - Sleep thread for up to 50
+    sem_post(&agentSem); //need new agent to provide more ingredients
+    printf("Smoker %d (tobacco-holding) smokes cigarette %d\n", num, i);
+    sleep((rand()%50)/1000);  //Smoke cigarette - Sleep thread for up to 50
+  }
+  printf("Smoker %d (tobacco-holding) is hungry, exits\n", num);
+}
+
+void *matchSmoker(void* arg){
+  int num = *((int*)arg); //unpack arg
+  for (int i = 0; i < 3; i++) {
+    sem_wait(&matchSem);
+    sleep((rand()%50)/1000); //Make cigarette - Sleep thread for up to 50
+    sem_post(&agentSem); //need new agent to provide more ingredients
+    printf("Smoker %d (match-holding) smokes cigarette %d\n", num, i);
+    sleep((rand()%50)/1000);  //Smoke cigarette - Sleep thread for up to 50
+  }
+  printf("Smoker %d (match-holding) is hungry, exits\n", num);
+}
+
+void *paperSmoker(void* arg){
+  int num = *((int*)arg);
+  for (int i = 0; i < 3; i++) {
+   sem_wait(&paperSem);
+   sleep((rand()%50)/1000); //Make cigarette - Sleep thread for up to 50
+   sem_post(&agentSem); //need new agent to provide more ingredients
+   printf("Smoker %d (paper-holding) smokes cigarette %d\n", num, i);
+   sleep((rand()%50)/1000); //Smoke cigarette - Sleep thread for up to 50
+  }
+  printf("Smoker %d (paper-holding) is hungry, exits\n", num);
+}
+
+int main(){
+  srand(time(NULL));   // Seed random times
+
+  /* initialize semiphores */
+  sem_init(&agentSem, 0, 1); // 1 agent at the table
+  // None of the smokers have supplies
+  sem_init(&tobaccoSem, 0, 0);
+  sem_init(&paperSem, 0, 0);
+  sem_init(&matchSem, 0, 0);
+  // No supplies on the table
+  sem_init(&tobacco, 0, 0);
+  sem_init(&paper, 0, 0);
+  sem_init(&match, 0, 0);
+  sem_init(&mutex, 0, 1);
+
+  //declare all pthreads
+  pthread_t smokers[6];
+  pthread_t paperPush, matchPush, tobaccoPush;
+  pthread_t agent1, agent2, agent3;
+
+  //create all pthreads - 3 agents, 3 pushers, 6 smokers
+  pthread_create(&agent1, NULL, agentOne, NULL);
+  pthread_create(&agent2, NULL, agentTwo, NULL);
+  pthread_create(&agent3, NULL, agentThree, NULL);
+  pthread_create(&paperPush, NULL, paperPusher, NULL);
+  pthread_create(&matchPush, NULL, matchPusher, NULL);
+  pthread_create(&tobaccoPush, NULL, tobaccoPusher, NULL);
+
+  // 6 smokers - 2 each of tobacco, matches, paper
+  int arg[6] = {0,1,2,3,4,5};
+  for (int i = 0; i < 6; i++)
+  {
+    if(i % 3 == 0)   pthread_create(&smokers[i], NULL, tobaccoSmoker, (void*)& arg[i]);
+    if(i % 3 == 1)   pthread_create(&smokers[i], NULL, matchSmoker, (void*)& arg[i]);
+    if(i % 3 == 2)   pthread_create(&smokers[i], NULL, paperSmoker, (void*)& arg[i]);
+  }
+
+  //terminate all pthreads
+  pthread_join(agent1, NULL);
+  pthread_join(agent2, NULL);
+  pthread_join(agent3, NULL);
+  pthread_join(paperPush, NULL);
+  pthread_join(matchPush, NULL);
+  pthread_join(tobaccoPush, NULL);
+  for (int i = 0; i < 6; i++) pthread_join(smokers[i], NULL);
+
+  printf("everybody's done!\n");
+
+  return 0;
+}
